requirements.txt
numpy
python-dotenv
google-generativeai